Wrap map in navigation container
