+++ index.js
@@ -1,25 +1,11 @@
 import { registerRootComponent } from 'expo';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-
-
-import { NavigationContainer } from '@react-navigation/native';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import App from './App';
-
-const Stack = createNativeStackNavigator();
 
 function Root() {
     return (
         <GestureHandlerRootView style={{ flex: 1 }}>
-            <NavigationContainer>
-                <Stack.Navigator
-                    screenOptions={{
-                        headerShown: false
-                    }}
-                >
-                    <Stack.Screen name="home" component={App} />
-                </Stack.Navigator>
-            </NavigationContainer>
+            <App />
         </GestureHandlerRootView>
     );
 }
